+++ src/git/remotes/gitlab.ts
@@ -3,11 +3,11 @@ import { GitHubService } from './github';
 
 export class GitLabService extends GitHubService {
 
-    constructor(public domain: string, public path: string) {
+    constructor(public domain: string, public path: string, public custom: boolean = false) {
         super(domain, path);
     }
 
     get name() {
-        return 'GitLab';
+        return this.formatName('GitLab');
     }
 }